pyproject.toml
[tool.poetry]
name = "wrtn"
version = "0.1.0"
description = "For wrtn"
authors = ["KIM SEUNG HOON"]

[tool.poetry.dependencies]
python = "^3.8"
mysql-connector-python = "^8.0.26"
playwright = "^1.38.0"
aiofiles = "^0.8.0"
requests = "^2.28.1"

[build-system]
requires = ["poetry>=1.0.0"]
build-backend = "poetry.masonry.api"